Implement verbose output option in data visualization
This commit includes the introduction of a `--verbose` command line
option to provide more detailed output.

The `scale_data` function is adjusted to generate verbose strings when
the `--verbose` option is active. For the "arrows" style, the function
will explain whether each data point has increased, decreased, or
remained the same as the previous data point. For the other styles, the
function will specify the value of each data point.

The `print_ansi_spark` function is also adjusted to print each string
from `scale_data` on a new line when the `--verbose` option is active.

The `argparse.ArgumentParser` in the `main` function is updated to
include the `--verbose` option.```